Refactored role configuration stuff into it's own place.
